requirements.txt
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
db-dtypes==1.0.4
google-api-core==2.10.2
google-auth==2.13.0
google-auth-oauthlib==0.7.0
google-cloud==0.34.0
google-cloud-bigquery==3.3.5
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
idna==3.4
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
pandas-gbq==0.17.9
proto-plus==1.22.1
protobuf==4.21.9
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.12
